refactor: split OAuth auth hooks by provider
